bot/requirements.txt
python-telegram-bot==20.7
anthropic==0.40.0
requests==2.31.0
python-dotenv==1.0.0
